+++ sentinelone/release_notes.md
@@ -1,6 +1,5 @@
 #### What's Improved
 
 - Added the following operations and playbooks:
-    - Fetch Threat File
-    - Get Threat Timeline
-    - Custom API Endpoint
+    - Create Blacklist Item
+    - Get Alerts